Resolve most of remaining circular dependencies
