+++ src/internal/index.ts
@@ -16,9 +16,19 @@ export * from './Configuration/ConfigurationTypes';
 
 export * from './Connection/ActionCode';
 export * from './Connection/CallbackLists';
-export * from './Connection/ConnectionTypes';
 export * from './Connection/ConnectionApi';
-export * from './Connection/MessageReceivers/index';
+export * from './Connection/ConnectionTypes';
+export * from './Connection/MessageReceivers/AnalyticsMessageReceiver';
+export * from './Connection/MessageReceivers/BaseMessageReceiver';
+export * from './Connection/MessageReceivers/ConfigStateMessageReceiver';
+export * from './Connection/MessageReceivers/HandDataHandler';
+export * from './Connection/MessageReceivers/HandPresenceMessageReceiver';
+export * from './Connection/MessageReceivers/InputActionMessageReceiver';
+export * from './Connection/MessageReceivers/InteractionZoneMessageReceiver';
+export * from './Connection/MessageReceivers/ResponseMessageReceiver';
+export * from './Connection/MessageReceivers/ServiceStateMessageReceiver';
+export * from './Connection/MessageReceivers/TrackingStateMessageReceiver';
+export * from './Connection/MessageReceivers/VersionHandshakeMessageReceiver';
 export * from './Connection/RequestTypes';
 export * from './Connection/ServiceConnection';
 export * from './Connection/ServiceTypes';
